tidy up metrics
