+++ birdset/train.py
@@ -67,7 +67,6 @@ def train(cfg):
         num_epochs=cfg.trainer.max_epochs, #?
         len_trainset=datamodule.len_trainset,
         batch_size=datamodule.loaders_config.train.batch_size,
-        label_counts=datamodule.num_train_labels,
     )
 
     object_dict = {
